Split context-menu for specific table actions
